fix(f52): tres consertos que o CI acusou, e um deles nao era assercao vermelha
1. cargo_dev 5e citava a assinatura ANTIGA de existe_outro_admin_ativo.
   has_function_privilege com assinatura inexistente LEVANTA em vez de devolver
   false — entao a asserçao nao ficava vermelha, ela ABORTAVA o roteiro inteiro
   e escondia tudo o que vinha depois.

2. import_fora_da_unidade 1a chamava prefixo_backup_import(1) com literal
   inteiro. int -> smallint no Postgres e conversao de ATRIBUICAO, nao
   implicita: resolve para 'function does not exist' e aborta. Cast explicito.

3. troca.sql faz um import DE VERDADE no caminho feliz, e foi escrito quando
   qualquer texto servia de backup. Ganhou a linha em storage.objects, o caminho
   sob o prefixo da filial e a chave confirmacao. Conserto de FIXTURE: o cenario
   1 continua testando que o substituto nasce por troca e nao por compra.

papeis_rls 3i nao precisou de conserto: ele roda como nao-admin e o e_admin()
da orquestradora recusa ANTES da resolucao da filial, entao nunca alcanca a
cascata nova.

+++ supabase/tests/import_fora_da_unidade.sql
@@ -116,11 +116,15 @@ begin
   -- =========================================================================
 
   -- 1a — concatenação pura, não depende de a filial 1 existir de verdade.
-  if public.prefixo_backup_import(1) = 'import/filial-1/' then
+  -- ⚠ O `::smallint` é OBRIGATÓRIO: a função recebe `smallint`, e a conversão de
+  -- `integer` para `smallint` no Postgres é de ATRIBUIÇÃO, não implícita — um literal
+  -- `1` cru resolve para "function public.prefixo_backup_import(integer) does not exist"
+  -- e ABORTA o roteiro, em vez de deixar a asserção vermelha.
+  if public.prefixo_backup_import(1::smallint) = 'import/filial-1/' then
     v_ok := v_ok + 1; raise notice '✓ 1a prefixo_backup_import(1) = "import/filial-1/"';
   else
     v_falhas := v_falhas + 1;
-    raise warning '✗ 1a prefixo_backup_import(1): esperado "import/filial-1/", obtido "%"', public.prefixo_backup_import(1);
+    raise warning '✗ 1a prefixo_backup_import(1): esperado "import/filial-1/", obtido "%"', public.prefixo_backup_import(1::smallint);
   end if;
 
   -- 1b — fechada nos quatro papéis, pelo ACL real (não pela intenção do comentário).
